feat(protocol): rename act -> task exports and migrate TaskId to task-id
