+++ packages/protocol/src/task/index.ts
@@ -1,2 +1,2 @@
-export * from "./act";
-export * from "./act-id";
+export * from "./task";
+export * from "./task-id";
